shinano-common: use mpdecision bin + ext lib from samsung klte
The stock Sony binary and extension library seem to not play well with
the QCOM PowerHAL, which seems to be modified by Sony on stock ROMs.
During operation, a "failed to acquire lock" warning will be spammed in
logcat everytime the screen is touched and the powerHAL fails to do
its job.

Use a binary and extension library from another msm8974 device to get
around this issue.

Change-Id: I09a744f965d7f770d9cbf63ffa5ce6f165cad7ee